updated readme,added travis.yml

+++ models/models.go
@@ -1,9 +1,9 @@
 package models
 
 // create models
-type User struct{
-	ID			int		`json:"id"`
-	Name		string	`json:"name"`
-	Age			uint8	`json:"age"`
-	Location	string	`json:"location"`
+type User struct {
+	ID       int64  `json:"id"`
+	Name     string `json:"name"`
+	Age      int64  `json:"age"`
+	Location string `json:"location"`
 }
